Prevent error inserting large MMSI numbers into the dirty table.
Error caused by large MMSIs (> maxint) being inserted into dirty table. In Python all int = long so these values are not truncated in the parsing stage, but will be rejected by the MMSI validation. We allow MMSIs in the dirty table to be bigint so such large values can still be inserted.